added a new wallpaper product item type, added a query for it, specifically modified newly created items to make appropriate Application/Wallpaper/Ringtone class validations before adding item to product catalog, and fixed a bug in Wallpaper where the pixel width/height was incorrectly returning a float when those properties are meant to be ints

+++ src/cscie97/asn2/ecommerce/product/Wallpaper.java
@@ -37,7 +37,7 @@ public class Wallpaper extends Content {
      *
      * @return  the width of the wallpaper in pixels
      */
-    public float getPixelWidth() {
+    public int getPixelWidth() {
         return pixelWidth;
     }
 
@@ -55,7 +55,7 @@ public class Wallpaper extends Content {
      *
      * @return  the height of the wallpaper in pixels
      */
-    public float getPixelHeight() {
+    public int getPixelHeight() {
         return pixelHeight;
     }
